.claude/skills/linestamp/requirements.txt
# LINEスタンプ生成スキル依存関係
google-genai>=1.0.0
Pillow>=10.0.0
PyYAML>=6.0.0
python-dotenv>=1.0.0

# Agent SDK（オプション）
claude-agent-sdk>=0.1.0

# 背景透過機能（オプション）
# rembg>=2.0.0
# onnxruntime-gpu>=1.18.0  # GPU使用時
# onnxruntime>=1.18.0      # CPU使用時
